refactor: case-insensitive path handling and update stimulus path resolution
- central `_ci_resolve`, `_ci_exists`, and `_ci_glob` in `data_path_utils`
- Replace redundant path-handling logic in `preflight`, `stimulus`, and `multipleye_data_collection`
