+++ preprocessing/data_collection/multipleye_data_collection.py
@@ -20,6 +20,7 @@ from tqdm import tqdm
 from ..models.sid import Sid
 from ..models.dcn import Dcn
 from ..config import settings
+from ..utils.data_path_utils import _ci_resolve
 from ..utils.conversion import convert_to_time_str
 from ..utils.logging import get_logger
 from ..checks.et_quality_checks import (
@@ -155,9 +156,9 @@ class MultipleyeDataCollection:
 
         # load stimulus order versions to know what stimulus randomization was used for
         # each participant
-        stim_order_versions = (
+        stim_order_versions = _ci_resolve(
             self.stimulus_dir / "config" / f"stimulus_order_versions_"
-            f"{self.language.casefold()}_{self.country.casefold()}_{self.lab_number}.csv"
+            f"{self.language}_{self.country}_{self.lab_number}.csv"
         )
         stim_order_versions = pd.read_csv(stim_order_versions)
         self.stim_order_versions = stim_order_versions[
